requirements.txt
jds6600>=0.6.0
pyserial>=3.5
